requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-huggingface
chromadb
beautifulsoup4
sentence-transformers
langchain_chroma
langchain.prompts
langchain_community